requirements.txt
imageio==2.8.0
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
matplotlib==3.2.1
opencv-python==4.2.0.32
Pillow==7.0.0
pytesseract==0.3.3
scikit-image==0.16.2
scipy==1.4.1
tensorflow==2.1.0
